ref(oauth): Extract shared authorization flow
Move OAuth link setup and private Slack delivery into a focused shared module so slash commands and callback handlers no longer depend on the jr-rpc shell surface.

This reduces interface bloat in jr-rpc, keeps plugin OAuth types with the plugin manifest types, and trims helper-only exports that were not part of the runtime contract.

Refs #75
