+++ packages/junior/src/chat/capabilities/jr-rpc-command.ts
@@ -1,4 +1,3 @@
-import { randomBytes } from "node:crypto";
 import { Bash, defineCommand } from "just-bash";
 import { listCapabilityProviders } from "@/chat/capabilities/catalog";
 import type { SkillCapabilityRuntime } from "@/chat/capabilities/runtime";
@@ -6,100 +5,10 @@ import { parseRepoTarget } from "@/chat/capabilities/target";
 import type { ChannelConfigurationService } from "@/chat/configuration/types";
 import type { UserTokenStore } from "@/chat/credentials/user-token-store";
 import { CredentialUnavailableError } from "@/chat/credentials/broker";
-import { logInfo, logWarn } from "@/chat/observability";
+import { formatProviderLabel, startOAuthFlow } from "@/chat/oauth-flow";
+import { logInfo } from "@/chat/observability";
 import { getPluginOAuthConfig } from "@/chat/plugins/registry";
-import { getSlackClient, isDmChannel } from "@/chat/slack-actions/client";
 import type { Skill } from "@/chat/skills";
-import { getStateAdapter } from "@/chat/state";
-
-/**
- * Deliver a private message to the requesting user. Authorization links must
- * ONLY be visible to the requesting user — never posted as visible channel
- * messages. See specs/security-policy.md "OAuth authorization link privacy".
- *
- * Delivery strategy:
- * 1. In 1:1 DMs (D-prefix): chat.postMessage (already private).
- * 2. In channels/groups: chat.postEphemeral (only the user sees it).
- * 3. If step 1 or 2 fails: open a DM with the user via conversations.open
- *    and send there.
- * 4. If all fail: return false. Caller must NOT expose the URL.
- */
-async function deliverPrivateMessage(input: {
-  channelId?: string;
-  threadTs?: string;
-  userId: string;
-  text: string;
-}): Promise<"in_context" | "fallback_dm" | false> {
-  let client: ReturnType<typeof getSlackClient>;
-  try {
-    client = getSlackClient();
-  } catch {
-    logWarn(
-      "oauth_private_delivery_skip",
-      {},
-      { "app.reason": "missing_bot_token" },
-      "Skipped private message delivery — no SLACK_BOT_TOKEN",
-    );
-    return false;
-  }
-
-  // Strategy 1 & 2: Try in-context delivery (ephemeral or DM message)
-  if (input.channelId) {
-    const isDm = isDmChannel(input.channelId);
-    try {
-      if (isDm) {
-        await client.chat.postMessage({
-          channel: input.channelId,
-          text: input.text,
-          ...(input.threadTs ? { thread_ts: input.threadTs } : {}),
-        });
-      } else {
-        await client.chat.postEphemeral({
-          channel: input.channelId,
-          user: input.userId,
-          text: input.text,
-          ...(input.threadTs ? { thread_ts: input.threadTs } : {}),
-        });
-      }
-      return "in_context";
-    } catch (error) {
-      const slackError = error instanceof Error ? error.message : String(error);
-      logWarn(
-        "oauth_private_delivery_failed",
-        {},
-        { "app.slack.error": slackError, "app.slack.channel": input.channelId },
-        `${isDm ? "DM" : "Ephemeral"} message delivery failed, falling back to DM`,
-      );
-    }
-  }
-
-  // Strategy 3: Open a DM with the user and send there
-  try {
-    const openResult = await client.conversations.open({ users: input.userId });
-    const dmChannelId = openResult.channel?.id;
-    if (!dmChannelId) {
-      logWarn(
-        "oauth_dm_fallback_failed",
-        {},
-        { "app.reason": "no_dm_channel_id" },
-        "conversations.open returned no channel ID",
-      );
-      return false;
-    }
-
-    await client.chat.postMessage({ channel: dmChannelId, text: input.text });
-    return "fallback_dm";
-  } catch (error) {
-    const slackError = error instanceof Error ? error.message : String(error);
-    logWarn(
-      "oauth_dm_fallback_failed",
-      {},
-      { "app.slack.error": slackError },
-      "DM fallback delivery failed",
-    );
-    return false;
-  }
-}
 
 type JrRpcDeps = {
   capabilityRuntime: SkillCapabilityRuntime;
@@ -219,7 +128,7 @@ async function handleIssueCredentialCommand(
     // Auto-start OAuth when no credentials are available for an OAuth-capable provider
     if (
       error instanceof CredentialUnavailableError &&
-      getOAuthProviderConfig(error.provider) &&
+      getPluginOAuthConfig(error.provider) &&
       deps.requesterId
     ) {
       const oauthResult = await startOAuthFlow(error.provider, {
@@ -231,8 +140,7 @@ async function handleIssueCredentialCommand(
         activeSkillName: deps.activeSkill?.name ?? undefined,
       });
       if (oauthResult.ok) {
-        const providerLabel =
-          error.provider.charAt(0).toUpperCase() + error.provider.slice(1);
+        const providerLabel = formatProviderLabel(error.provider);
         return commandResult({
           stdout: {
             credential_unavailable: true,
@@ -459,147 +367,6 @@ function isKnownProvider(provider: string): boolean {
   return listCapabilityProviders().some((p) => p.provider === provider);
 }
 
-export type OAuthProviderConfig = {
-  clientIdEnv: string;
-  clientSecretEnv: string;
-  authorizeEndpoint: string;
-  tokenEndpoint: string;
-  scope?: string;
-  authorizeParams?: Record<string, string>;
-  tokenAuthMethod?: "body" | "basic";
-  tokenExtraHeaders?: Record<string, string>;
-  callbackPath: string;
-};
-
-export function getOAuthProviderConfig(
-  provider: string,
-): OAuthProviderConfig | undefined {
-  return getPluginOAuthConfig(provider);
-}
-
-export type OAuthStatePayload = {
-  userId: string;
-  provider: string;
-  channelId?: string;
-  threadTs?: string;
-  pendingMessage?: string;
-  configuration?: Record<string, unknown>;
-};
-
-const OAUTH_STATE_TTL_MS = 10 * 60 * 1000; // 10 minutes
-
-export function resolveBaseUrl(): string | undefined {
-  const explicit = process.env.JUNIOR_BASE_URL?.trim();
-  if (explicit) return explicit;
-  const vercelProd = process.env.VERCEL_PROJECT_PRODUCTION_URL?.trim();
-  if (vercelProd) return `https://${vercelProd}`;
-  const vercelUrl = process.env.VERCEL_URL?.trim();
-  if (vercelUrl) return `https://${vercelUrl}`;
-  return undefined;
-}
-
-export type OAuthFlowInput = {
-  requesterId: string;
-  channelId?: string;
-  threadTs?: string;
-  userMessage?: string;
-  channelConfiguration?: ChannelConfigurationService;
-  activeSkillName?: string;
-};
-
-export async function startOAuthFlow(
-  provider: string,
-  input: OAuthFlowInput,
-): Promise<
-  | { ok: false; error: string }
-  | { ok: true; delivery: "in_context" | "fallback_dm" | false }
-> {
-  const providerConfig = getOAuthProviderConfig(provider);
-  if (!providerConfig) {
-    return {
-      ok: false,
-      error: `Provider "${provider}" does not support OAuth authorization`,
-    };
-  }
-
-  const clientId = process.env[providerConfig.clientIdEnv]?.trim();
-  if (!clientId) {
-    return {
-      ok: false,
-      error: `Missing ${providerConfig.clientIdEnv} environment variable`,
-    };
-  }
-
-  const baseUrl = resolveBaseUrl();
-  if (!baseUrl) {
-    return {
-      ok: false,
-      error:
-        "Cannot determine base URL (set JUNIOR_BASE_URL or deploy to Vercel)",
-    };
-  }
-
-  // Snapshot channel configuration so the resumed turn has context
-  let configuration: Record<string, unknown> | undefined;
-  if (input.userMessage && input.channelConfiguration) {
-    configuration = await input.channelConfiguration.resolveValues();
-  }
-
-  const state = randomBytes(32).toString("hex");
-  const stateKey = `oauth-state:${state}`;
-  const stateAdapter = getStateAdapter();
-  const statePayload: OAuthStatePayload = {
-    userId: input.requesterId,
-    provider,
-    ...(input.channelId ? { channelId: input.channelId } : {}),
-    ...(input.threadTs ? { threadTs: input.threadTs } : {}),
-    ...(input.userMessage ? { pendingMessage: input.userMessage } : {}),
-    ...(configuration && Object.keys(configuration).length > 0
-      ? { configuration }
-      : {}),
-  };
-  await stateAdapter.set(stateKey, statePayload, OAUTH_STATE_TTL_MS);
-
-  const redirectUri = `${baseUrl}${providerConfig.callbackPath}`;
-  const params = new URLSearchParams({
-    client_id: clientId,
-    state,
-    redirect_uri: redirectUri,
-    response_type: "code",
-  });
-  if (providerConfig.scope) {
-    params.set("scope", providerConfig.scope);
-  }
-  for (const [key, value] of Object.entries(
-    providerConfig.authorizeParams ?? {},
-  )) {
-    params.set(key, value);
-  }
-  const authorizeUrl = `${providerConfig.authorizeEndpoint}?${params.toString()}`;
-
-  logInfo(
-    "jr_rpc_oauth_start",
-    {},
-    {
-      "app.credential.provider": provider,
-      ...(input.activeSkillName
-        ? { "app.skill.name": input.activeSkillName }
-        : {}),
-    },
-    "Initiated OAuth authorization code flow",
-  );
-
-  const providerLabel = provider.charAt(0).toUpperCase() + provider.slice(1);
-  const delivery = await deliverPrivateMessage({
-    channelId: input.channelId,
-    threadTs: input.threadTs,
-    userId: input.requesterId,
-    text: `<${authorizeUrl}|Click here to link your ${providerLabel} account>. Once you've authorized, you'll see a confirmation in Slack.`,
-  });
-
-  return { ok: true, delivery };
-}
-
 async function handleOAuthStartCommand(
   args: string[],
   deps: JrRpcDeps,
@@ -626,8 +393,7 @@ async function handleOAuthStartCommand(
       stored &&
       (stored.expiresAt === undefined || stored.expiresAt > Date.now())
     ) {
-      const providerLabel =
-        provider.charAt(0).toUpperCase() + provider.slice(1);
+      const providerLabel = formatProviderLabel(provider);
       return commandResult({
         stdout: {
           ok: true,
